Removed the method tabCompleteRegions from WorldGuard7 and WorldGuard6_1 classes since the method signature and implementation are identical, and created the same method in the super type WorldGuardInterface class.
