+++ arminterface/src/main/java/net/alex9849/inter/WorldGuardInterface.java
@@ -5,11 +5,14 @@ import com.sk89q.worldguard.protection.flags.InvalidFlagFormat;
 import com.sk89q.worldguard.protection.flags.RegionGroup;
 import com.sk89q.worldguard.protection.flags.RegionGroupFlag;
 import com.sk89q.worldguard.protection.managers.RegionManager;
+import com.sk89q.worldguard.protection.regions.ProtectedRegion;
 import org.bukkit.Location;
 import org.bukkit.World;
 import org.bukkit.entity.Player;
 
+import java.util.ArrayList;
 import java.util.List;
+import java.util.Map;
 
 public abstract class WorldGuardInterface {
 
@@ -31,7 +34,16 @@ public abstract class WorldGuardInterface {
 
     public abstract RegionGroup parseFlagInput(RegionGroupFlag flag, String input) throws InvalidFlagFormat;
 
-    public abstract List<String> tabCompleteRegions(String regionName, World world);
+    public List<String> tabCompleteRegions(String regionName, World world) {
+        Map<String, ProtectedRegion> regions = this.getRegionManager(world).getRegions();
+        List<String> regionIds = new ArrayList<>();
+        for(Map.Entry<String, ProtectedRegion> entry : regions.entrySet()) {
+            if(entry.getKey().toLowerCase().startsWith(regionName.toLowerCase())) {
+                regionIds.add(entry.getValue().getId());
+            }
+        }
+        return regionIds;
+    }
 
     public abstract Flag fuzzyMatchFlag(String id);
 }
