+++ wg6_1adapter/src/main/java/net/alex9849/adapters/WorldGuard6_1.java
@@ -82,16 +82,4 @@ public class WorldGuard6_1 extends WorldGuardInterface {
         return flag.parseInput(WorldGuardPlugin.inst(), null, input);
     }
 
-    @Override
-    public List<String> tabCompleteRegions(String regionName, World world) {
-        Map<String, ProtectedRegion> regions = this.getRegionManager(world).getRegions();
-        List<String> regionIds = new ArrayList<>();
-        for(Map.Entry<String, ProtectedRegion> entry : regions.entrySet()) {
-            if(entry.getKey().toLowerCase().startsWith(regionName.toLowerCase())) {
-                regionIds.add(entry.getValue().getId());
-            }
-        }
-        return regionIds;
-    }
-
 }
